Seperate hadith schemas and topic schemas

+++ src/data/models/topic/topic-translation.ts
@@ -3,10 +3,8 @@ import { TranslationSchema } from "../base/base"
 import { TopicTranslatedFieldsSchema } from "./topic"
 
 export const TopicTranslationSchema = TranslationSchema.extend({
-  topicID: z.number().int(),
+  topicID: z.number(),
   ...TopicTranslatedFieldsSchema.shape,
 })
-
 type TopicTranslation = z.infer<typeof TopicTranslationSchema>
-
 export interface ITopicTranslation extends TopicTranslation {}
